Change variable type in AKG function from int to ll for better precision

+++ test.cpp
@@ -26,7 +26,7 @@ const int MOD = 1e9+7;
 ll N = 1e18;
 
 void AKG(){
-    int a,b,c,x,y;cin>>a>>b>>c>>x>>y;
+    ll a,b,c,x,y;cin>>a>>b>>c>>x>>y;
 	if(a>=x&&b>=y)yn(1);
 	else if(((x-a)+(y-b))>=c)yn(1);
 	else yn(0);
